Move serverless state to folder
