+++ packages/twenty-front/src/modules/serverless-functions/components/ServerlessFunctionExecutionResult.tsx
@@ -1,8 +1,8 @@
 import {
-  type ExecutionStatus,
-  WorkflowStepExecutionResult,
+    type ExecutionStatus,
+    WorkflowStepExecutionResult,
 } from '@/workflow/components/WorkflowStepExecutionResult';
-import { type ServerlessFunctionTestData } from '@/workflow/states/serverlessFunctionTestDataFamilyState';
+import { type ServerlessFunctionTestData } from '@/workflow/workflow-steps/workflow-actions/code-action/states/serverlessFunctionTestDataFamilyState';
 import { ServerlessFunctionExecutionStatus } from '~/generated-metadata/graphql';
 
 export const ServerlessFunctionExecutionResult = ({
